Reorganize fields, make fields lazy
