+++ edgy/core/db/fields/__init__.py
@@ -1,41 +1,46 @@
+from typing import TYPE_CHECKING
+
+from monkay import Monkay
+
 from .base import BaseField
-from .composite_field import CompositeField
-from .core import (
-    AutoNowMixin,
-    BigIntegerField,
-    BinaryField,
-    BooleanField,
-    CharField,
-    ChoiceField,
-    ComputedField,
-    DateField,
-    DateTimeField,
-    DecimalField,
-    DurationField,
-    EmailField,
-    FloatField,
-    IntegerField,
-    IPAddressField,
-    JSONField,
-    PasswordField,
-    PlaceholderField,
-    SmallIntegerField,
-    TextField,
-    TimeField,
-    URLField,
-    UUIDField,
-)
-from .exclude_field import ExcludeField
-from .file_field import FileField
-from .foreign_keys import ForeignKey
-from .image_field import ImageField
-from .many_to_many import ManyToMany, ManyToManyField
-from .one_to_one_keys import OneToOne, OneToOneField
-from .ref_foreign_key import RefForeignKey
 from .types import BaseFieldType
 
+if TYPE_CHECKING:
+    from .composite_field import CompositeField
+    from .core import (
+        AutoNowMixin,  # noqa
+        BigIntegerField,
+        BinaryField,
+        BooleanField,
+        CharField,
+        ChoiceField,
+        ComputedField,
+        DateField,
+        DateTimeField,
+        DecimalField,
+        DurationField,
+        EmailField,
+        FloatField,
+        IntegerField,
+        IPAddressField,
+        JSONField,
+        PasswordField,
+        PlaceholderField,
+        SmallIntegerField,
+        TextField,
+        TimeField,
+        URLField,
+        UUIDField,
+    )
+    from .exclude_field import ExcludeField
+    from .file_field import FileField
+    from .foreign_keys import ForeignKey
+    from .image_field import ImageField
+    from .many_to_many import ManyToMany, ManyToManyField
+    from .one_to_one_keys import OneToOne, OneToOneField
+    from .ref_foreign_key import RefForeignKey
+
 __all__ = [
-    "AutoNowMixin",
     "BaseField",
     "BaseFieldType",
     "BigIntegerField",
@@ -71,3 +76,49 @@ __all__ = [
     "URLField",
     "UUIDField",
 ]
+
+Monkay(
+    globals(),
+    lazy_imports={
+        "BigIntegerField": ".core.BigIntegerField",
+        "BinaryField": ".core.BinaryField",
+        "BooleanField": ".core.BooleanField",
+        "CharField": ".core.CharField",
+        "ChoiceField": ".core.ChoiceField",
+        "ComputedField": ".computed_field.ComputedField",
+        "CompositeField": ".composite_field.CompositeField",
+        "DateField": ".core.DateField",
+        "DateTimeField": ".core.DateTimeField",
+        "DurationField": ".core.DurationField",
+        "DecimalField": ".core.DecimalField",
+        "EmailField": ".core.EmailField",
+        "ExcludeField": ".exclude_field.ExcludeField",
+        "FloatField": ".core.FloatField",
+        "FileField": ".file_field.FileField",
+        "ImageField": ".image_field.ImageField",
+        "ForeignKey": ".foreign_keys.ForeignKey",
+        "IntegerField": ".core.IntegerField",
+        "IPAddressField": ".core.IPAddressField",
+        "JSONField": ".core.JSONField",
+        "RefForeignKey": ".ref_foreign_key.RefForeignKey",
+        "ManyToMany": ".many_to_many.ManyToMany",
+        "ManyToManyField": ".many_to_many.ManyToMany",
+        "OneToOne": ".one_to_one_keys.OneToOne",
+        "OneToOneField": ".one_to_one_keys.OneToOne",
+        "PasswordField": ".core.PasswordField",
+        "PlaceholderField": ".place_holder_field.PlaceholderField",
+        "SmallIntegerField": ".core.SmallIntegerField",
+        "TextField": ".core.TextField",
+        "TimeField": ".core.TimeField",
+        "URLField": ".core.URLField",
+        "UUIDField": ".core.UUIDField",
+    },
+    deprecated_lazy_imports={
+        "AutoNowMixin": {
+            "path": ".mixins.AutoNowMixin",
+            "reason": "We export mixins now from edgy.core.db.fields.mixins.",
+            "new_attribute": "edgy.core.db.fields.mixins.AutoNowMixin",
+        }
+    },
+    skip_all_update=True,
+)
